AI_Models/requirements_training.txt
# ─── Kaggle Training ───────────────────────────────────────────────────────────────────────
# Detection & Segmentation
ultralytics>=8.0.0
segmentation-models>=1.0.1

# Deep Learning frameworks
tensorflow>=2.12.0
torch>=2.0.0
torchvision>=0.15.0

# NLP - ClinicalBERT (HuggingFace)
transformers>=4.38.0
tokenizers>=0.15.0
sentencepiece>=0.1.99

# Machine Learning Ensemble
xgboost>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
joblib>=1.3.0

# XAI
shap>=0.44.0

# Data Processing
opencv-python>=4.8.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
